Initial work on reworking the depth averaging.
Oops, add the files

astyle
