+++ source/simulator/helper_functions.cc
@@ -881,404 +881,6 @@ namespace aspect
       }
   }
 
-
-  template <int dim>
-  template<class FUNCTOR>
-  void Simulator<dim>::compute_depth_average(std::vector<double> &values,
-                                             FUNCTOR &fctr) const
-  {
-    Assert (values.size() > 0,
-            ExcMessage ("To call this function, you need to request a positive "
-                        "number of depth slices."));
-    const unsigned int num_slices = values.size();
-    std::vector<double> volume(num_slices);
-
-    // this yields 10^dim quadrature points evenly distributed in the interior of the cell.
-    // We avoid points on the faces, as they would be counted more than once.
-    const QIterated<dim> quadrature_formula (QMidpoint<1>(),
-                                             10);
-    const unsigned int n_q_points = quadrature_formula.size();
-    const double max_depth = geometry_model->maximal_depth();
-
-    FEValues<dim> fe_values (mapping,
-                             finite_element,
-                             quadrature_formula,
-                             update_values | update_gradients | update_quadrature_points);
-
-    std::vector<std::vector<double> > composition_values (parameters.n_compositional_fields,std::vector<double> (n_q_points));
-
-    std::vector<double> output_values(quadrature_formula.size());
-
-    typename DoFHandler<dim>::active_cell_iterator
-    cell = dof_handler.begin_active(),
-    endc = dof_handler.end();
-
-    MaterialModel::MaterialModelInputs<dim> in(n_q_points,
-                                               parameters.n_compositional_fields);
-    MaterialModel::MaterialModelOutputs<dim> out(n_q_points,
-                                                 parameters.n_compositional_fields);
-
-    fctr.setup(quadrature_formula.size());
-
-    for (; cell!=endc; ++cell)
-      if (cell->is_locally_owned())
-        {
-          fe_values.reinit (cell);
-          if (fctr.need_material_properties())
-            {
-              fe_values[introspection.extractors.pressure].get_function_values (this->solution,
-                                                                                in.pressure);
-              fe_values[introspection.extractors.temperature].get_function_values (this->solution,
-                                                                                   in.temperature);
-              fe_values[introspection.extractors.velocities].get_function_values (this->solution,
-                                                                                  in.velocity);
-              fe_values[introspection.extractors.velocities].get_function_symmetric_gradients (this->solution,
-                  in.strain_rate);
-              fe_values[introspection.extractors.pressure].get_function_gradients (this->solution,
-                                                                                   in.pressure_gradient);
-              for (unsigned int c=0; c<parameters.n_compositional_fields; ++c)
-                fe_values[introspection.extractors.compositional_fields[c]].get_function_values(this->solution,
-                                                                                                composition_values[c]);
-
-              for (unsigned int i=0; i<n_q_points; ++i)
-                {
-                  in.position[i] = fe_values.quadrature_point(i);
-                  for (unsigned int c=0; c<parameters.n_compositional_fields; ++c)
-                    in.composition[i][c] = composition_values[c][i];
-                }
-              in.cell = &cell;
-
-              material_model->evaluate(in, out);
-            }
-
-          fctr(in, out, fe_values, this->solution, output_values);
-
-          for (unsigned int q = 0; q < n_q_points; ++q)
-            {
-              const double depth = geometry_model->depth(fe_values.quadrature_point(q));
-              const unsigned int idx = static_cast<unsigned int>((depth*num_slices)/max_depth);
-              Assert(idx<num_slices, ExcInternalError());
-
-              values[idx] += output_values[q] * fe_values.JxW(q);
-              volume[idx] += fe_values.JxW(q);
-            }
-        }
-
-    std::vector<double> values_all(num_slices);
-    std::vector<double> volume_all(num_slices);
-    Utilities::MPI::sum(volume, mpi_communicator, volume_all);
-    Utilities::MPI::sum(values, mpi_communicator, values_all);
-
-    for (unsigned int i=0; i<num_slices; ++i)
-      values[i] = values_all[i] / (static_cast<double>(volume_all[i])+1e-20);
-  }
-
-  namespace
-  {
-    template <int dim>
-    class FunctorDepthAverageField
-    {
-      public:
-        FunctorDepthAverageField(const FEValuesExtractors::Scalar &field)
-          : field_(field) {}
-
-        bool need_material_properties() const
-        {
-          return false;
-        }
-
-        void setup(const unsigned int)
-        {
-        }
-
-        void operator()(const MaterialModel::MaterialModelInputs<dim> &,
-                        const MaterialModel::MaterialModelOutputs<dim> &,
-                        FEValues<dim> &fe_values,
-                        const LinearAlgebra::BlockVector &solution,
-                        std::vector<double> &output)
-        {
-          fe_values[field_].get_function_values (solution, output);
-        }
-
-        const FEValuesExtractors::Scalar field_;
-    };
-  }
-
-  template <int dim>
-  void Simulator<dim>::compute_depth_average_field(const AdvectionField &advection_field,
-                                                   std::vector<double> &values) const
-  {
-    const FEValuesExtractors::Scalar field
-      = (advection_field.is_temperature()
-         ?
-         introspection.extractors.temperature
-         :
-         introspection.extractors.compositional_fields[advection_field.compositional_variable]
-        );
-
-    FunctorDepthAverageField<dim> f(field);
-    compute_depth_average(values, f);
-  }
-
-  namespace
-  {
-    template <int dim>
-    class FunctorDepthAverageViscosity
-    {
-      public:
-        bool need_material_properties() const
-        {
-          return true;
-        }
-
-        void setup(const unsigned int)
-        {}
-
-        void operator()(const MaterialModel::MaterialModelInputs<dim> &,
-                        const MaterialModel::MaterialModelOutputs<dim> &out,
-                        FEValues<dim> &,
-                        const LinearAlgebra::BlockVector &,
-                        std::vector<double> &output)
-        {
-          output = out.viscosities;
-        }
-    };
-  }
-
-  template <int dim>
-  void Simulator<dim>::compute_depth_average_viscosity(std::vector<double> &values) const
-  {
-    FunctorDepthAverageViscosity<dim> f;
-    compute_depth_average(values, f);
-  }
-
-
-  namespace
-  {
-    template <int dim>
-    class FunctorDepthAverageVelocityMagnitude
-    {
-      public:
-        FunctorDepthAverageVelocityMagnitude(const FEValuesExtractors::Vector &field)
-          : field_(field) {}
-
-        bool need_material_properties() const
-        {
-          return false;
-        }
-
-        void setup(const unsigned int q_points)
-        {
-          velocity_values.resize(q_points);
-        }
-
-        void operator()(const MaterialModel::MaterialModelInputs<dim> &,
-                        const MaterialModel::MaterialModelOutputs<dim> &,
-                        FEValues<dim> &fe_values,
-                        const LinearAlgebra::BlockVector &solution,
-                        std::vector<double> &output)
-        {
-          fe_values[field_].get_function_values (solution, velocity_values);
-          for (unsigned int q=0; q<output.size(); ++q)
-            output[q] = velocity_values[q] * velocity_values[q];
-        }
-
-        std::vector<Tensor<1,dim> > velocity_values;
-        const FEValuesExtractors::Vector field_;
-    };
-  }
-
-
-  template <int dim>
-  void Simulator<dim>::compute_depth_average_velocity_magnitude(std::vector<double> &values) const
-  {
-    FunctorDepthAverageVelocityMagnitude<dim> f(introspection.extractors.velocities);
-
-    compute_depth_average(values, f);
-  }
-
-  namespace
-  {
-    template <int dim>
-    class FunctorDepthAverageSinkingVelocity
-    {
-      public:
-        FunctorDepthAverageSinkingVelocity(const FEValuesExtractors::Vector &field, GravityModel::Interface<dim> *gravity)
-          : field_(field), gravity_(gravity) {}
-
-        bool need_material_properties() const
-        {
-          // this is needed because we want to access in.position in operator()
-          return true;
-        }
-
-        void setup(const unsigned int q_points)
-        {
-          velocity_values.resize(q_points);
-        }
-
-        void operator()(const MaterialModel::MaterialModelInputs<dim> &in,
-                        const MaterialModel::MaterialModelOutputs<dim> &,
-                        FEValues<dim> &fe_values,
-                        const LinearAlgebra::BlockVector &solution,
-                        std::vector<double> &output)
-        {
-          fe_values[field_].get_function_values (solution, velocity_values);
-          for (unsigned int q=0; q<output.size(); ++q)
-            {
-              const Tensor<1,dim> g = gravity_->gravity_vector(in.position[q]);
-              const Tensor<1,dim> vertical = (g.norm() > 0 ? g/g.norm() : Tensor<1,dim>());
-
-              output[q] = std::fabs(std::min(0.0, velocity_values[q] * vertical))*year_in_seconds;
-            }
-        }
-
-        std::vector<Tensor<1,dim> > velocity_values;
-        const FEValuesExtractors::Vector field_;
-        GravityModel::Interface<dim> *gravity_;
-    };
-  }
-
-
-  template <int dim>
-  void Simulator<dim>::compute_depth_average_sinking_velocity(std::vector<double> &values) const
-  {
-    FunctorDepthAverageSinkingVelocity<dim> f(introspection.extractors.velocities,
-                                              this->gravity_model.get());
-
-    compute_depth_average(values, f);
-  }
-
-  namespace
-  {
-    template <int dim>
-    class FunctorDepthAverageVsVp
-    {
-      public:
-        FunctorDepthAverageVsVp(const MaterialModel::Interface<dim> *mm, bool vs)
-          : material_model(mm), vs_(vs)
-        {}
-
-        bool need_material_properties() const
-        {
-          return true;
-        }
-
-        void setup(const unsigned int)
-        {}
-
-        void operator()(const MaterialModel::MaterialModelInputs<dim> &in,
-                        const MaterialModel::MaterialModelOutputs<dim> &,
-                        FEValues<dim> &,
-                        const LinearAlgebra::BlockVector &,
-                        std::vector<double> &output)
-        {
-          if (vs_)
-            for (unsigned int q=0; q<output.size(); ++q)
-              output[q] = material_model->seismic_Vs(
-                            in.temperature[q], in.pressure[q], in.composition[q],
-                            in.position[q]);
-          else
-            for (unsigned int q=0; q<output.size(); ++q)
-              output[q] = material_model->seismic_Vp(
-                            in.temperature[q], in.pressure[q], in.composition[q],
-                            in.position[q]);
-        }
-
-        const MaterialModel::Interface<dim> *material_model;
-        bool vs_;
-    };
-
-  }
-
-  template <int dim>
-  void Simulator<dim>::compute_depth_average_Vs(std::vector<double> &values) const
-  {
-    FunctorDepthAverageVsVp<dim> f(this->material_model.get(), true /* Vs */);
-
-    compute_depth_average(values, f);
-  }
-
-  template <int dim>
-  void Simulator<dim>::compute_depth_average_Vp(std::vector<double> &values) const
-  {
-    FunctorDepthAverageVsVp<dim> f(this->material_model.get(), false /* Vp */);
-
-    compute_depth_average(values, f);
-  }
-
-  namespace
-  {
-    template <int dim>
-    class FunctorDepthAverageVerticalHeatFlux
-    {
-      public:
-        FunctorDepthAverageVerticalHeatFlux(const FEValuesExtractors::Vector &velocity_field,
-                                            const FEValuesExtractors::Scalar &temperature_field,
-                                            const GravityModel::Interface<dim> *gm)
-          : velocity_field_(velocity_field),
-            temperature_field_(temperature_field),
-            gravity_model(gm)
-        {}
-
-        bool need_material_properties() const
-        {
-          return true;
-        }
-
-        void setup(const unsigned int q_points)
-        {
-          velocity_values.resize(q_points);
-          temperature_values.resize(q_points);
-          temperature_gradients.resize(q_points);
-        }
-
-        void operator()(const MaterialModel::MaterialModelInputs<dim> &in,
-                        const MaterialModel::MaterialModelOutputs<dim> &out,
-                        FEValues<dim> &fe_values,
-                        const LinearAlgebra::BlockVector &solution,
-                        std::vector<double> &output)
-        {
-          fe_values[velocity_field_].get_function_values (solution, velocity_values);
-          fe_values[temperature_field_].get_function_values (solution, temperature_values);
-          fe_values[temperature_field_].get_function_gradients (solution, temperature_gradients);
-
-          for (unsigned int q=0; q<output.size(); ++q)
-            {
-              const Tensor<1,dim> gravity = gravity_model->gravity_vector(in.position[q]);
-              const Tensor<1,dim> vertical = -gravity/( gravity.norm() != 0.0 ?
-                                                        gravity.norm() : 1.0 );
-              const double advective_flux = (velocity_values[q] * vertical) * in.temperature[q] *
-                                            out.densities[q]*out.specific_heat[q];
-              const double conductive_flux = -(temperature_gradients[q]*vertical) *
-                                             out.thermal_conductivities[q];
-              output[q] = advective_flux + conductive_flux;
-            }
-        }
-
-        const FEValuesExtractors::Vector velocity_field_;
-        const FEValuesExtractors::Scalar temperature_field_;
-        const GravityModel::Interface<dim> *gravity_model;
-        std::vector<Tensor<1,dim> > velocity_values;
-        std::vector<Tensor<1,dim> > temperature_gradients;
-        std::vector<double> temperature_values;
-    };
-
-  }
-
-  template <int dim>
-  void Simulator<dim>::compute_depth_average_vertical_heat_flux(std::vector<double> &values) const
-  {
-    FunctorDepthAverageVerticalHeatFlux<dim> f(introspection.extractors.velocities,
-                                               introspection.extractors.temperature,
-                                               this->gravity_model.get());
-
-    compute_depth_average(values, f);
-  }
-
-
-
-
   template <int dim>
   bool
   Simulator<dim>::stokes_matrix_depends_on_solution() const
@@ -1303,13 +905,6 @@ namespace aspect
   template std::pair<double,double> Simulator<dim>::get_extrapolated_advection_field_range (const AdvectionField &advection_field) const; \
   template std::pair<double,bool> Simulator<dim>::compute_time_step () const; \
   template void Simulator<dim>::make_pressure_rhs_compatible(LinearAlgebra::BlockVector &vector); \
-  template void Simulator<dim>::compute_depth_average_field(const AdvectionField &advection_field, std::vector<double> &values) const; \
-  template void Simulator<dim>::compute_depth_average_viscosity(std::vector<double> &values) const; \
-  template void Simulator<dim>::compute_depth_average_velocity_magnitude(std::vector<double> &values) const; \
-  template void Simulator<dim>::compute_depth_average_sinking_velocity(std::vector<double> &values) const; \
-  template void Simulator<dim>::compute_depth_average_Vs(std::vector<double> &values) const; \
-  template void Simulator<dim>::compute_depth_average_Vp(std::vector<double> &values) const; \
-  template void Simulator<dim>::compute_depth_average_vertical_heat_flux(std::vector<double> &values) const; \
   template void Simulator<dim>::output_program_stats(); \
   template void Simulator<dim>::output_statistics(); \
   template double Simulator<dim>::compute_initial_stokes_residual(); \
